examples/python/requirements.txt
PyJWT
hashlib
websocket-client